pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vulyk"
version = "0.6.0"
description = "Crowdsourcing platform for different kinds of tasks"
readme = { file = "README.rst", content-type = "text/x-rst" }
license = { text = "BSD 3-Clause License" }
authors = [
    { name = "Dmytro Hambal", email = "mr_hambal@outlook.com" },
    { name = "Dmitry Chaplinsky", email = "chaplinsky.dmitry@gmail.com" },
    { name = "Volodymyr Hotsyk", email = "gotsyk@gmail.com" },
    { name = "Alexander Sapronov" },

]
keywords = ["collaboration", "flask", "crowdsourcing", "python", "vulyk"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "blinker>1.4,<2.0.0",
    "bz2file==0.98",
    "click>8.1.3",
    "rcssmin==1.2.1",
    "email-validator==2.2.0",
    "flask>2.0.0,<3.0.0",
    "flask-admin==1.6.1",
    "flask-assets==2.1.0",
    "Flask-Collect-Invenio==1.4.0",
    "flask-login==0.6.3",
    "flask-mongoengine-3>1.0.0",
    "flask-script==2.0.6",
    "flask-wtf~=1.2.0",
    "honcho>=0.5.0",
    "itsdangerous>=2.2",
    "jinja2>=3.0",
    "markupsafe>2.1.1",
    "orjson~=3.9",
    "pillow>=5.4.0",
    "rjsmin==1.2.4",
    "social-auth-app-flask-mongoengine<2.0.0",
    "social-auth-app-flask<2.0.0",
    "social-auth-core==4.5.6",
    "prettytable~=3.16.0",
    "webassets==2.0",
    "werkzeug>3.0",
]

[dependency-groups]
dev = [
    "coverage > 7",
    "gunicorn==23.0.0",
    "ruff >=0.11.0,<1.0.0",
    "mypy >=1.15.0,<2.0.0",
    "sphinx>7.0,<8",
    "types-WTForms",
#    "vulyk-declaration",
]

#[tool.uv.sources]
#vulyk-declaration = { path = "../vulyk-declaration" }

[project.urls]
Homepage = "https://github.com/mrgambal/vulyk"
Documentation = "https://vulyk.readthedocs.org"
Repository = "https://github.com/mrgambal/vulyk"

[project.scripts]
manage = "manage:main"

[tool.mypy]
exclude = ['./tests']
python_version = "3.13"
show_error_codes = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = false

[tool.ruff]
src = ["./vulyk", "tests"]
line-length = 120
output-format = "pylint"
target-version = "py310"

[tool.ruff.lint]
select = [
    "B",      # flakes8 bugbear
    "E",      # pycodestyle errors
    "F",      # flake8 pyflakes
    "Q",      # flake8-quotes
    "W",      # pycodestyle warnings
    "C90",    # flake8-complexity / xenon
    "N",      # pep8-naming
    "S",      # flake8-bandit
    "BLE",    # flake8-blind-except
    "FBT",    # flake8-boolean-trap
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimeez
    "I",      # flake8-import-order
    "ISC003", # implicit-str-concat
    "PGH",    # pygrep hooks
    "PIE",    # flake8-pie
    "PYI",    # flake8-pyi
    "RUF",    # ruff-specific
    "RET",    # flake8-return
    "SLF",    # flake8-self
    "SIM",    # flake8-simplify
    "T20",    # flake8-print
    "YTT",    # flake8-2020
    "FURB",   # refurb
]
ignore = [
    "N818", # error-suffix-on-exception-name
]

[tool.ruff.lint.isort]
known-first-party = ["vulyk"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"

[tool.ruff.lint.per-file-ignores]
# Ignore `RUF009` (Do not perform function call `os.getenv` in dataclass defaults) in all `config.py` files.
"config.py" = ["RUF009"]
"**/test_*.py" = ["S101", "SLF001"]

[tool.hatch.build.targets.wheel]
include = ["vulyk/**/*"]

[tool.hatch.build]
include = [
    "vulyk/**/*.py",
    "vulyk/**/*.html",
    "vulyk/**/*.css",
    "vulyk/**/*.js",
    "vulyk/**/*.png",
    "vulyk/**/*.jpg",
    "vulyk/**/*.gif",
    "README.rst",
    "HISTORY.rst",
    "LICENSE",
]
packages = ["vulyk"]

[tool.uv]
package = false
